requirements.txt
dj-database-url==0.4.2
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==2.0.6
Django==1.10.5
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
PyJWT==1.4.2
